requirements.txt
py-cord[voice]>=2.6.0
openai-whisper>=20231117
pynacl>=1.5.0
python-dotenv>=1.0.0
